[refactor] dto에서 model로 변경
