+++ data/src/main/java/com/prac/data/source/api/GitHubApi.kt
@@ -1,14 +1,12 @@
 package com.prac.data.source.api
 
-import com.prac.data.source.model.RepoModel
-import com.prac.data.source.model.TokenModel
+import com.prac.data.source.dto.RepoDto
 import retrofit2.http.GET
-import retrofit2.http.Header
 import retrofit2.http.Path
 
 internal interface GitHubApi {
     @GET("users/{userName}/repos")
     suspend fun getRepos(
         @Path("userName") userName: String
-    ): List<RepoModel>
+    ): List<RepoDto>
 }
