+++ data/src/main/java/com/prac/data/source/dto/OwnerDto.kt
@@ -1,14 +1,14 @@
-package com.prac.data.source.model
+package com.prac.data.source.dto
 
 import com.prac.data.repository.dto.OwnerDto
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-internal data class OwnerModel(
+internal data class OwnerDto(
     @SerialName("login") val login: String,
     @SerialName("avatar_url") val avatarUrl: String
 ) {
-    fun toDto() =
+    fun toModel() =
         OwnerDto(login, avatarUrl)
 }
