+++ data/src/main/java/com/prac/data/source/dto/RepoDto.kt
@@ -1,17 +1,17 @@
-package com.prac.data.source.model
+package com.prac.data.source.dto
 
 import com.prac.data.repository.dto.RepoDto
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-internal data class RepoModel(
+internal data class RepoDto(
     @SerialName("id") val id: Int,
     @SerialName("name") val name: String,
-    @SerialName("owner") val owner: OwnerModel,
+    @SerialName("owner") val owner: OwnerDto,
     @SerialName("stargazers_count") val stargazersCount: Int,
     @SerialName("updated_at") val updatedAt: String
 ) {
-    fun toDto() =
-        RepoDto(id, name, owner.toDto(), stargazersCount, updatedAt)
+    fun toModel() =
+        RepoDto(id, name, owner.toModel(), stargazersCount, updatedAt)
 }
