+++ data/src/main/java/com/prac/data/source/dto/TokenDto.kt
@@ -1,10 +1,10 @@
-package com.prac.data.source.model
+package com.prac.data.source.dto
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-internal data class TokenModel(
+internal data class TokenDto(
     @SerialName("access_token") val accessToken: String,
     @SerialName("expires_in") val expiresIn: Int,
     @SerialName("refresh_token") val refreshToken: String,
